refactor: useValidation 구조 변경
